Done working on a dashboard functionlaity after login and done working on a clickable dashboard when login from Trainer. as well as done working on a Athlete Dashboard API.

+++ SportsBallistics/app/src/main/java/com/example/sportsballistics/data/remote/athletes/AthleteDataModel.kt
@@ -15,9 +15,9 @@ data class AthleteDataModel(
 
 data class Service(
     val name: String,
-    val average: Int,
-    val sum: Int,
-    val percent: Int,
+    val average: Double,
+    val sum: Double,
+    val percent: Double,
     val slug: String,
     val slug_edit: String,
 )
